requirements.txt
boto==2.49.0
boto3==1.21.28
botocore==1.24.28
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.7
django-cors-headers==3.11.0
django-heroku==0.3.1
django-pwa==1.0.10
django-storages==1.12.3
djangorestframework==3.13.1
gunicorn==20.1.0
mysql-connector==2.2.9
mysqlclient==2.1.0
Pillow==8.3.2
sqlparse==0.4.2
static3==0.7.0
storage==0.0.4.3
whitenoise==5.3.0

